// src/features/comment/api/models/view/comment.view.model.ts
import { LikeStatus } from '../../../../../infrastructure/utils/enums';

export type CommentViewModel = {
  id: string;
  content: string;
  createdAt: string;
  commentatorInfo: {
    userId: string;
    userLogin: string;
  };
  likesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: LikeStatus;
  };
};